exercise 4.21

+++ part4/controllers/blogs.js
@@ -43,10 +43,19 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response
-        .status(204)
-        .end()
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response
+            .status(204)
+            .end()
+    } else {
+        return response.status(401).json({error: 'only blog owner can delete it'})
+    }
+
+    
 })
 
 blogsRouter.put('/:id', async (request, response) => {
